Stufe 3.3 Weg B: Features + ChannelOptions als Wert-Typen aus DTO baubar
Erkenntnis der Bestandsaufnahme: diese Aggregat-Zweige sind reine Wert-Typen
mit Fachsemantik (Feature-Defaults, InteractiveGUIAccess-Regel,
modify()-Arithmetik) ohne Laufzeit-Kopplung. Schnitt daher einfacher als das
Sicht-Interface-Muster: Wert-Typ bleibt erhalten (Semantik nur an einer
Stelle, Konsumenten unberuehrt), nur der Creator faellt spaeter; der Mapper
baut ueber neue Factories (Features.of, AllChannelOptions.of ->
Mapper.toFeatures/toChannelOptions). ChannelDto auf nullable Bindung
umgestellt (Double fuer fix/factor/offset wie der alte Parser - modify()
unterscheidet nicht-gesetzt von 0; powerOnDelay_s als Integer mit
x1000-Ableitung und Default -1 in Mapper.powerOnDelayMs). Tests: Features aus
Domaene UND DTO in allen 11 Abfragen + Map identisch (Template +
Minimal-Fixture inkl. per-Feature-Defaults); Regelverletzung wirft auch am
DTO-Weg; ChannelOption-Liste charakterisiert. Offen notiert: benannte
Feature-Alt-Elemente (XSD-choice) noch ungebunden. 65 Tests gruen.

+++ src/de/sgollmer/solvismax/xml/jaxb/BaseDataDto.java
@@ -117,16 +117,21 @@ public class BaseDataDto {
 
 	/**
 	 * JAXB-Bean für {@code <Channel>} — Kanal-Sonderoption. Pro Eintrag ist i. d. R.
-	 * nur eines von {@code fix}/{@code offset}/{@code powerOnDelay_s} gesetzt; die
-	 * nicht gesetzten int-Attribute sind kanonisch {@code 0}.
+	 * nur eines von {@code fix}/{@code offset}/{@code powerOnDelay_s} gesetzt.
+	 * Nullable Wrapper-Typen ({@code null} = Attribut fehlt), weil der alte
+	 * Parser {@code fix}/{@code factor}/{@code offset} als {@code Double} mit
+	 * {@code null}-Semantik bindet — {@code ChannelOption.modify()} unterscheidet
+	 * „nicht gesetzt" von {@code 0}. {@code powerOnDelay_s}: Sekunden-Rohwert,
+	 * die ×1000-Ableitung (Default −1 bei fehlendem Attribut) liegt in
+	 * {@code Mapper.powerOnDelayMs}.
 	 */
 	@XmlAccessorType(XmlAccessType.FIELD)
 	public static class ChannelDto {
 		@XmlAttribute(name = "id") public String id;
-		@XmlAttribute(name = "fix") public int fix;
-		@XmlAttribute(name = "factor") public int factor;
-		@XmlAttribute(name = "offset") public int offset;
-		@XmlAttribute(name = "powerOnDelay_s") public int powerOnDelay_s;
+		@XmlAttribute(name = "fix") public Double fix;
+		@XmlAttribute(name = "factor") public Double factor;
+		@XmlAttribute(name = "offset") public Double offset;
+		@XmlAttribute(name = "powerOnDelay_s") public Integer powerOnDelay_s;
 	}
 
 	/** JAXB-Bean für {@code <Features>}. */
